+++ jest.config.js
@@ -1,7 +1,5 @@
-/** @type {import('@ts-jest/dist/types').InitialOptionsTsJest} */
-module.exports = {
+export default {
   preset: 'ts-jest',
   testEnvironment: 'jsdom',
-  collectCoverage: true,
   coverageDirectory: 'coverage',
 }