+++ src/misc/event.ts
@@ -1,8 +1,10 @@
+import type { Fn } from './types'
+
 export const noop = () => {}
 
 export function on<T extends Window | Document | HTMLElement | EventTarget>(
   obj: T | null,
-  ...args: Parameters<T['addEventListener']> | [string, Function | null, ...any]
+  ...args: Parameters<T['addEventListener']> | [string, Fn | null, ...any]
 ): void {
   if (obj && obj.addEventListener) {
     obj.addEventListener(...(args as Parameters<HTMLElement['addEventListener']>))
@@ -11,7 +13,7 @@ export function on<T extends Window | Document | HTMLElement | EventTarget>(
 
 export function off<T extends Window | Document | HTMLElement | EventTarget>(
   obj: T | null,
-  ...args: Parameters<T['removeEventListener']> | [string, Function | null, ...any]
+  ...args: Parameters<T['removeEventListener']> | [string, Fn | null, ...any]
 ): void {
   if (obj && obj.removeEventListener) {
     obj.removeEventListener(...(args as Parameters<HTMLElement['removeEventListener']>))