+++ CHANGELOG.md
@@ -2,6 +2,19 @@
 
 All notable changes to this project will be documented in this file.
 
+### v0.1.4 (2024-08-15)
+
+#### Changed
+
+- Replaced `yarn` with `pnpm`.
+- Upgraded `React` and `react-dom` peerDependencies versions.
+- Upgraded ESLint to v9.
+- Updated unit test scripts.
+
+#### Removed
+
+- Removed `CommonJS` module.
+
 ### v0.1.2 (2021-07-27)
 
 - fix tsconfig (remove helper).